pyproject.toml
[project]
name = "data-science"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.12",
    "langgraph>=0.4.1",
    "mlflow>=2.22.0",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "torch>=2.7.0",
]
